split views and models into different directories
