+++ splitbackend/views/__init__.py
@@ -1,0 +1,13 @@
+# from .tab_views import (
+#     ParticipatedTabList,
+#     OwnedTabList,
+#     CreateTab,
+#     TabDetail,
+# )
+# from .user_views import (
+#     Hello,
+#     user_create,
+#     ProfileList
+# )
+from .tab_views import *
+from .user_views import *
